+++ ksyun/provider.go
@@ -38,9 +38,11 @@ func Provider() terraform.ResourceProvider {
 			"ksyun_lines":                         dataSourceKsyunLines(),
 			"ksyun_eips":                          dataSourceKsyunEips(),
 			"ksyun_slbs":                          dataSourceKsyunSlbs(),
+			"ksyun_lbs":                           dataSourceKsyunLbs(),
 			"ksyun_listeners":                     dataSourceKsyunListeners(),
 			"ksyun_health_checks":                 dataSourceKsyunHealthChecks(),
 			"ksyun_listener_servers":              dataSourceKsyunListenerServers(),
+			"ksyun_lb_listener_servers":           dataSourceKsyunLbListenerServers(),
 			"ksyun_lb_acls":                       dataSourceKsyunSlbAcls(),
 			"ksyun_availability_zones":            dataSourceKsyunAvailabilityZones(),
 			"ksyun_network_interfaces":            dataSourceKsyunNetworkInterfaces(),
